+++ power/decimate/__init__.py
@@ -1,0 +1,4 @@
+from ._decimate import (
+	make_decimated_power.
+	m_pxy_decimated
+	)