+++ power/decimate/__main__.py
@@ -1,0 +1,6 @@
+import sys
+
+from ._decimate import make_decimated_power
+
+simdir = sys.argv[1]
+make_decimated_power(simdir)